Move film render method and handlers from PageController to FilmController
